refactor(cache): separate adapter conformance checks
